feat(teebag): move SGX DCAP types to dedicated module
Move all DCAP quote and report related types to a new module to improve code
organization and readability. This includes the DcapQuote, SgxReportBody,
SgxQuote and related structs. No functional changes, just code reorganization.

Types were moved from sgx_verify/mod.rs to the new dcap_types.rs module and
re-exported through teebag/mod.rs to maintain compatibility.
